Major refactoring: Poller replaced by Notifier
POller has been replaced by Notifier which returns a channel of events
that can be consumed and acted upon.
