+++ runtime/runtime.go
@@ -1,6 +1,8 @@
 // Package runtime is a service runtime manager
 package runtime
 
+import "time"
+
 var (
 	// DefaultRuntime is default micro runtime
 	DefaultRuntime = newRuntime()
@@ -22,6 +24,52 @@ type Runtime interface {
 	Stop() error
 }
 
+// Notifier is an update notifier
+type Notifier interface {
+	// Notify publishes notification events
+	Notify() (<-chan Event, error)
+	// Close stops the notifier
+	Close() error
+}
+
+// EventType defines notification event
+type EventType int
+
+const (
+	// Create is emitted when a new build has been craeted
+	Create EventType = iota
+	// Update is emitted when a new update become available
+	Update
+	// Delete is emitted when a build has been deleted
+	Delete
+)
+
+// String returns human readable event type
+func (t EventType) String() string {
+	switch t {
+	case Create:
+		return "create"
+	case Delete:
+		return "delete"
+	case Update:
+		return "update"
+	default:
+		return "unknown"
+	}
+}
+
+// Event is notification event
+type Event struct {
+	// Type is event type
+	Type EventType
+	// Timestamp is event timestamp
+	Timestamp time.Time
+	// Service is the name of the service
+	Service string
+	// Version of the build
+	Version string
+}
+
 // Service is runtime service
 type Service struct {
 	// Name of the service
@@ -34,46 +82,4 @@ type Service struct {
 	Exec string
 	// Version of the service
 	Version string
-}
-
-func Create(s *Service, opts ...CreateOption) error {
-	return DefaultRuntime.Create(s, opts...)
-}
-
-func Delete(s *Service) error {
-	return DefaultRuntime.Delete(s)
-}
-
-func Update(s *Service) error {
-	return DefaultRuntime.Update(s)
-}
-
-func List() ([]*Service, error) {
-	return DefaultRuntime.List()
-}
-
-func Start() error {
-	return DefaultRuntime.Start()
-}
-
-func Stop() error {
-	return DefaultRuntime.Stop()
-}
-
-// NewRuntime returns new runtime
-func NewRuntime(opts ...Option) Runtime {
-	options := Options{}
-	// apply requested options
-	for _, o := range opts {
-		o(&options)
-	}
-
-	switch options.Type {
-	case "local":
-		return newRuntime(opts...)
-	case "k8s":
-		return NewK8sRuntime(opts...)
-	default:
-		return newRuntime(opts...)
-	}
 }
